Restructure system_api module
