+++ rs/embedders/fuzz/src/imports.rs
@@ -65,7 +65,7 @@ pub(crate) fn system_api_imports(config: EmbeddersConfig) -> SystemApiImportStor
 
     match config.feature_flags.wasm64 {
         FlagStatus::Enabled => {
-            system_api::syscalls::<u64>(
+            system_api::linker::syscalls::<u64>(
                 &mut linker,
                 config.feature_flags,
                 config.stable_memory_dirty_page_limit,
@@ -74,7 +74,7 @@ pub(crate) fn system_api_imports(config: EmbeddersConfig) -> SystemApiImportStor
             );
         }
         FlagStatus::Disabled => {
-            system_api::syscalls::<u32>(
+            system_api::linker::syscalls::<u32>(
                 &mut linker,
                 config.feature_flags,
                 config.stable_memory_dirty_page_limit,
